refactor: move query-building logic from domain to business-logic
Keep fetch-flows.ts as a thin API call. Move FetchFlowsParams,
DEFAULT_FLOWS_SORT, STATUS_CATEGORY_TO_BACKEND, and query assembly
into flows-queries.ts. Rename flowsQueries.all → flowsQueries.list.
