+++ src/modules/flows/business-logic/flows-queries.ts
@@ -1,10 +1,55 @@
 import { queryOptions } from "@tanstack/react-query";
-import {
-	DEFAULT_FLOWS_SORT,
-	type FetchFlowsParams,
-	fetchFlows,
-} from "../domain/fetch-flows";
+import { fetchFlows } from "../domain/fetch-flows";
 import { fetchFlow } from "../domain/fetch-flow";
+import type { FlowStatusFilter } from "../domain/flow";
+import {
+	FLOW_ACTIVE_STATUSES,
+	FLOW_COMPLETED_STATUSES,
+	FLOW_FAILED_STATUSES,
+} from "./categorize-flow-status";
+
+export const DEFAULT_FLOWS_SORT = "desc:latest_run";
+
+export type FetchFlowsParams = {
+	name?: string;
+	status?: FlowStatusFilter;
+	sort?: string;
+};
+
+const STATUS_CATEGORY_TO_BACKEND: Record<
+	Exclude<FlowStatusFilter, "all">,
+	readonly string[]
+> = {
+	running: FLOW_ACTIVE_STATUSES,
+	failed: FLOW_FAILED_STATUSES,
+	completed: FLOW_COMPLETED_STATUSES,
+};
+
+function buildFlowsQuery(params: FetchFlowsParams): Record<string, unknown> {
+	const query: Record<string, unknown> = {
+		sort_by: params.sort ?? DEFAULT_FLOWS_SORT,
+	};
+
+	let filterCount = 0;
+
+	const trimmedName = params.name?.trim();
+	if (trimmedName) {
+		query.name = `contains:${trimmedName}`;
+		filterCount++;
+	}
+
+	if (params.status && params.status !== "all") {
+		const backendValues = STATUS_CATEGORY_TO_BACKEND[params.status];
+		query.latest_run_status = `oneof:${JSON.stringify(backendValues)}`;
+		filterCount++;
+	}
+
+	if (filterCount > 1) {
+		query.logical_operator = "and";
+	}
+
+	return query;
+}
 
 export const flowsQueryKeys = {
 	all: ["flows"] as const,
@@ -26,11 +71,11 @@ function normalizeParams(params: FetchFlowsParams): FetchFlowsParams {
 }
 
 export const flowsQueries = {
-	all: (params: FetchFlowsParams = {}) => {
+	list: (params: FetchFlowsParams = {}) => {
 		const normalized = normalizeParams(params);
 		return queryOptions({
 			queryKey: flowsQueryKeys.list(normalized),
-			queryFn: () => fetchFlows(normalized),
+			queryFn: () => fetchFlows(buildFlowsQuery(normalized)),
 		});
 	},
 	detail: (flowId: string) =>
